:sparkles: feat : 수익률 계산 및 출력 기능 추가

+++ src/main/java/lotto/domain/Money.java
@@ -1,13 +1,13 @@
 package lotto.domain;
 
 public class Money {
-    private final int value;
+    private final long value;
 
-    public Money(final int value) {
+    public Money(final long value) {
         this.value = value;
     }
 
-    public int getValue() {
+    public long getValue() {
         return value;
     }
 }
